% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABHgenotypeR.R
\docType{package}
\name{ABHgenotypeR}
\alias{ABHgenotypeR}
\alias{ABHgenotypeR-package}
\title{ABHgenotypeR: A package for easy visualization of ABH genotypes.}
\description{
The ABHgenotypeR packages is meat as a companion packages in between the TASSEL
GBS pipeline and R/qtl. It allows easy visualization of ABH-encoded genotypes
in a .csv format as ouptut by the TASSEL ABHGenosPlugin which is also the
format necessary for R/qtl input.
}
\section{ABHgenotypeR functions}{
 The readHapmap... plot functions can then be
  utilized...
}

