% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeByFlanks.R
\name{imputeByFlanks}
\alias{imputeByFlanks}
\title{Impute missing genotypes based on flanking alleles}
\usage{
imputeByFlanks(inputGenos = "genotypes")
}
\arguments{
\item{inputGenos}{A genotypes list object.}
}
\value{
A genotype object in which missing data is imputed based on flanking
  alleles. Any number of N is replaced by either A, B or N if the alleles which flank the N match
}
\description{
Impute missing genotypes based on flanking alleles
}
\examples{
\dontrun{imputedGenos <- imputeByFlanks(genotypes)}
}

