% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlleleFreq.R
\name{plotAlleleFreq}
\alias{plotAlleleFreq}
\title{Plot the  parental allele frequencies along the chromosomes.}
\usage{
plotAlleleFreq(genos = "genotypes")
}
\arguments{
\item{genos}{The output of readABHgenotypes}
}
\value{
A plot of parental allele frequencies along the chromosomes. If the
  output is assigned a name a ggplot2 object is returned for further
  manipulation.
}
\description{
Plot the  parental allele frequencies along the chromosomes.
}
\examples{
plotAlleleFreq(genotypes)
p <- plotAlleleFreq(genotypes)
}

