% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMarkerDensity.R
\name{plotMarkerDensity}
\alias{plotMarkerDensity}
\title{Plot the marker density along the chromosomes.}
\usage{
plotMarkerDensity(genos = "genotypes")
}
\arguments{
\item{genos}{The output of readABHgenotypes}
}
\value{
A plot of marker densities along the chromosomes. If the output is
  assigned a name a ggplot2 object is returned for further manipulation.
}
\description{
Plot the marker density along the chromosomes.
}
\examples{
\dontrun{plotMarkerDensity(genotypes)}
\dontrun{p <- plotMarkerDensity(genotypes)}
}

