% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenos.R
\name{reportGenos}
\alias{reportGenos}
\title{Report the total and relative number of each allele in a genotype object.}
\usage{
reportGenos(genos1)
}
\arguments{
\item{genos1}{A genotypes list object.}
}
\value{
Console output of the total an relative number of each allele.
}
\description{
Report the total and relative number of each allele in a genotype object.
}
\examples{
\dontrun{reportGenos(preImputation)}
}

